requirements.txt
fastapi
uvicorn
playwright
python-dotenv
slowapi
pydantic
httpx
python-multipart